requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.18
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.1
aiohttp==3.8.4
python-dotenv==1.0.0

